+++ crawler/src/instances/garbage.rs
@@ -1,4 +1,0 @@
-include!(concat!(
-    env!("CARGO_MANIFEST_DIR"),
-    "/../../apps/garbage/crawler/src/instances/garbage.rs"
-));